Renames Group By node to Reduce By, changes name in code as well (was still named Sum), refactors code minorly to make more sense, small lint fixes
