+++ ui/client/components/aiAssistant/AIAssistant.js
@@ -84,7 +84,7 @@ const AIAssistant = () => {
         // TODO: Use this for local development, causemos-analyst for production
         // const queryResp = await axios.get(`http://localhost:8001/mock-message?query=${searchPhrase}`);
 
-        const mock = ''; //'mock-';
+        const mock = ''; // 'mock-';
 
         const queryResp = await axios.get(
           `/api/ai-docs/${mock}message?query=${searchPhrase}`
@@ -99,11 +99,11 @@ const AIAssistant = () => {
         // go through the unique filenames and fetch the document data
         const documentFetches = Object.keys(filenamesToDocs).map(async (filename) => {
           try {
-          const docFetchResp = await axios.get(`/api/dojo/documents/by-didx-name?name=${filename}`);
+            const docFetchResp = await axios.get(`/api/dojo/documents/by-didx-name?name=${filename}`);
             return { filename, data: docFetchResp.data };
           } catch (e) {
             console.log('error fetching document, skipping', e);
-            return {filename: null, data: {}};
+            return { filename: null, data: {} };
           }
         });
