+++ ui/client/components/aiAssistant/AssistantChatCard.js
@@ -1,7 +1,5 @@
 import React, { useState } from 'react';
 
-import AppBar from '@mui/material/AppBar';
-import Toolbar from '@mui/material/Toolbar';
 import IconButton from '@mui/material/IconButton';
 import CloseIcon from '@mui/icons-material/Close';
 import Button from '@mui/material/Button';
