--- ui/client/dagpipes/nodeLabels.js
@@ -1,9 +1,0 @@
-export default {
-  SUM: 'Group By',
-  SUM_DIMENSION: 'Dimension',
-  LOAD: 'Load',
-  SAVE: 'Save',
-  MULTIPLY: 'Join', // change example: 'Join'
-  FILTER_BY_COUNTRY: 'Filter By Country', // was Country Split
-  THRESHOLD: 'Threshold'
-};
